% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hazards.r
\name{add_hazards}
\alias{add_hazards}
\title{Add Hazards}
\usage{
add_hazards(dist1, dist2, ...)
}
\arguments{
\item{dist1}{survival distribution to add}

\item{dist2}{second survival distribution to add}

\item{...}{additional survival distributions to add}
}
\value{
a \code{surv_add_haz} object
}
\description{
Combine two or more survival distributions as independent risks
by adding their hazards.
}
\examples{

dist1 <- define_surv_param("exp", rate = .125)
dist2 <- define_surv_param("weibull", shape = 1.2, scale = 50)
combined_dist <- add_hazards(dist1, dist2)

}
