% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_af.r
\name{apply_af}
\alias{apply_af}
\title{Apply Acceleration Factor}
\usage{
apply_af(dist, af, log_af = FALSE)
}
\arguments{
\item{dist}{a survival distribution}

\item{af}{an acceleration factor to be applied to survival distribution}

\item{log_af}{optional argument (defaults to \code{FALSE}) to indicate that
provided acceleration factor is on log scale}
}
\value{
A \code{surv_aft} object.
}
\description{
Apply an acceleration factor to proportionally increase or reduce
the time-to-event of a survival distribution.
}
\examples{

dist1 <- define_surv_param("exp", rate = 0.25)
aft_dist <- apply_af(dist1, 1.5)
}
