% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_hr.r
\name{apply_hr}
\alias{apply_hr}
\title{Apply Hazard Ratio}
\usage{
apply_hr(dist, hr, log_hr = FALSE)
}
\arguments{
\item{dist}{a survival distribution}

\item{hr}{a hazard ratio to be applied to survival distribution}

\item{log_hr}{optional argument (defaults to \code{FALSE}) to indicate that provided
hazard ratio is on log scale}
}
\value{
A \code{surv_ph} object.
}
\description{
Apply hazard ratio to a distribution to proportionally reduce or increase
the hazard rate.
}
\examples{

ph_dist <- apply_hr(
 define_surv_param("exp", rate = 0.25),
 0.5
)

}
