% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_or.r
\name{apply_or}
\alias{apply_or}
\title{Apply Odds Ratio}
\usage{
apply_or(dist, or, log_or = FALSE)
}
\arguments{
\item{dist}{a survival distribution}

\item{or}{an odds ratio to be applied to survival distribution}

\item{log_or}{optional argument (defaults to \code{FALSE}) to indicate that
provided odds ratio is on log scale}
}
\value{
A \code{surv_po} object.
}
\description{
Apply an odds ratio to proportionally increase or reduce
the odds of survival
}
\examples{

dist1 <- define_surv_param("exp", rate = 0.25)
po_dist <- apply_or(dist1, 1.12)
}
