% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_shift.r
\name{apply_shift}
\alias{apply_shift}
\title{Apply Fixed Shift in Time}
\usage{
apply_shift(dist, shift)
}
\arguments{
\item{dist}{a survival distribution}

\item{shift}{amount of time to shift distribution, where a positive
value moves hazards forward and a negative value delays hazards.}
}
\value{
A \code{surv_shift} object.
}
\description{
Apply a fixed shift in time to move the hazards of a survival
distribution forwards or backwards.
}
\examples{

shifted_dist <- apply_shift(
 define_surv_param("exp", rate = 0.25),
 3
)

}
