% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_function.r
\name{define_function_surv}
\alias{define_function_surv}
\title{Define a survival distribution based on a function}
\usage{
define_function_surv(f, ...)
}
\arguments{
\item{f}{a function that takes a vector of times and returns a vector
of corresponding survival probabilities}

\item{...}{additional arguments to be passed to f}
}
\value{
a \code{surv_function} object.
}
\description{
Define a survival distribution based on a function
}
