% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_km.r
\name{define_surv_km}
\alias{define_surv_km}
\title{Define Kaplan-Meier Distribution}
\usage{
define_surv_km(x, time_col = "time", surv_col = "survival")
}
\arguments{
\item{x}{a data frame with columns for time and survival probability. By
default, these columns are assumed to be named \code{time} and \code{survival}, but
these can be configured by via the optional parameters.}

\item{time_col}{the name of the time column (defaults to \code{time})}

\item{surv_col}{the name of the time column (defaults to \code{survival})}
}
\value{
a \code{surv_km} object.
}
\description{
Define a survival distribution based on a table of Kaplan-Meier
output containing times and survival probabilities.
}
\examples{
df <- data.frame(
     time = c(0, 1, 5, 10),
     survival = c(1, 0.9, 0.7, 0.5)
)
define_surv_km(df)
 
}
