% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.r
\name{define_survival}
\alias{define_survival}
\title{Define a Survival Distribution}
\usage{
define_survival(distribution, ...)
}
\arguments{
\item{distribution}{A parametric survival distribution.}

\item{...}{Additional distribution parameters (see
respective distribution help pages).}
}
\value{
A \code{surv_dist} object.
}
\description{
Define a parametric survival distribution with given
parameter values.
}
\examples{

define_survival(distribution = "exp", rate = .5)
define_survival(distribution = "gompertz", rate = .5, shape = 1)

}
