% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.r
\name{event_prob}
\alias{event_prob}
\title{Evaluate Event Probabilities}
\usage{
event_prob(x, start, end, ...)
}
\arguments{
\item{x}{A \code{surv_dist} object}

\item{start}{A numeric vector of interval start times}

\item{end}{A numeric vector of interval end times}

\item{...}{additional arguments passed to methods}
}
\description{
Generate the conditional probability of an even during an
interval of time.
}
\examples{
dist1 <- define_surv_param('exp', rate = 0.12)
surv_prob(dist1, c(0, 1, 2, 3))
  
}
