% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_flexsurv.r
\name{extract_flexsurv_params}
\alias{extract_flexsurv_params}
\title{Extract Evaluated Parameters}
\usage{
extract_flexsurv_params(obj, data = NULL)
}
\arguments{
\item{obj}{A \code{\link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg()}} object.}

\item{data}{An optional dataset of covariate values to
generate parameters for. Defaults to the original data
to which the model was fit.}
}
\value{
A tidy data frame of curve parameters for each
covariate level.
}
\description{
Extracts the covariate-adjusted parameters from a
\code{\link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg()}} object.
}
