% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.r
\name{join}
\alias{join}
\title{Join Distributions}
\usage{
join(dist1, cut1, dist2, ...)
}
\arguments{
\item{dist1}{survival distribution to use from time \code{0} to \code{cut}}

\item{cut1}{cut point between \code{dist1} and \code{dist2}}

\item{dist2}{survival distribution to use from \code{cut}}

\item{...}{Additional cutpoints and distributions}
}
\value{
A \code{surv_join} object
}
\description{
Join two or more distributions together at the specified cut points.
}
\examples{

dist1 <- define_survival(distribution = "exp", rate = 0.05)
dist2 <- define_survival(distribution = "gompertz", rate = .5, shape = 1)
dist3 <- define_survival(distribution = "exp", rate = 0.25)
join_dist <- join(dist1, 20, dist2)
join_dist2 <- join(dist1, 20, dist2, 50, dist3)

}
