% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.r
\name{mix}
\alias{mix}
\title{Mix Two or More Survival Distributions}
\usage{
mix(dist1, weight1, dist2, weight2, ...)
}
\arguments{
\item{dist1}{first survival distribution to mix}

\item{weight1}{probability weight for first distribution}

\item{dist2}{second survival distribution to mix}

\item{weight2}{probability weight for second distribution}

\item{...}{additional distributions and weights}
}
\value{
A \code{surv_mix} object.
}
\description{
Mix a set of survival distributions using the specified
weights.
}
\examples{

dist1 <- define_surv_param("exp", rate = .5)
dist2 <- define_surv_param("gompertz", rate = .5, shape = 1)
pooled_dist <- mix(dist1, 0.25, dist2, 0.75)

}
