% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_covariates.r
\name{set_covariates}
\alias{set_covariates}
\title{Set Covariates of a Survival Model}
\usage{
set_covariates(dist, data, ...)
}
\arguments{
\item{dist}{a survfit or flexsurvreg object}

\item{data}{a data.frame representing subjets for which predictions
will be generated}

\item{...}{optional argument representing covariate values to generate
predictions for, can be used instead of data argument}
}
\value{
a \code{surv_model} object
}
\description{
Generate a survival distribution representing model predictions
for a specified cohort. The cohort can be defined by providing a
data frame of covariate values (for multiple subjects) or by providing
covariate values as named arguments (for a single subject).
}
\examples{

fs1 <- flexsurvreg(
  Surv(rectime, censrec)~group,
  data=flexsurv::bc,
  dist = "llogis"
)
good_model <- set_covariates(fs1, group = "Good")
cohort <- data.frame(group=c("Good", "Good", "Medium", "Poor"))
mixed_model <- set_covariates(fs1, data = cohort)
}
