% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_max.r
\name{surv_max}
\alias{surv_max}
\title{Get the Maximum of Two or More Survival Distributions}
\usage{
surv_max(dist1, dist2, ...)
}
\arguments{
\item{dist1}{first survival distribution}

\item{dist2}{second survival distribution}

\item{...}{additional distributions}
}
\value{
A \code{surv_max} object.
}
\description{
Finds the maximum of n provided survival distributions
}
