% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_hr.r
\name{apply_hr}
\alias{apply_hr}
\title{Apply a Hazard Ratio}
\usage{
apply_hr(dist, hr, log_hr = FALSE)
}
\arguments{
\item{dist}{A survival distribution.}

\item{hr}{A hazard ratio to be applied.}

\item{log_hr}{If \code{TRUE}, the hazard ratio is exponentiated
before being applied.}
}
\value{
A \code{surv_ph} object.
}
\description{
Proportionally reduce or increase the hazard rate of a
distribution.
}
\examples{

ph_dist <- apply_hr(
 define_surv_param("exp", rate = 0.25),
 0.5
)

}
