% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatability.r, R/generics.r, R/surv_cure.r,
%   R/surv_function.r, R/surv_parametric.r, R/surv_spline.r
\name{eval_surv}
\alias{eval_surv}
\alias{surv_prob}
\alias{surv_prob.surv_cure}
\alias{surv_prob.surv_function}
\alias{surv_prob.surv_parametric}
\alias{surv_prob.surv_spline}
\title{Evaluate Survival Probabilities}
\usage{
eval_surv(x, time, ...)

surv_prob(x, time, ...)

\method{surv_prob}{surv_cure}(x, time, ...)

\method{surv_prob}{surv_function}(x, time, ...)

\method{surv_prob}{surv_parametric}(x, time, ...)

\method{surv_prob}{surv_spline}(x, time, ...)
}
\arguments{
\item{x}{A surv_dist object}

\item{time}{A numeric vector of times}

\item{...}{additional arguments passed to methods.}
}
\value{
A numeric vector of survival probabilities
}
\description{
Generate survival probabilities for a survival distribution
at the specified times.
}
