% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.r
\name{surv_prob}
\alias{surv_prob}
\alias{eval_surv}
\title{Evaluate Survival Probabilities}
\usage{
surv_prob(x, time, ...)

eval_surv(x, time, ...)
}
\arguments{
\item{x}{A surv_dist object}

\item{time}{A numeric vector of times}

\item{...}{additional arguments passed to methods.}
}
\value{
A numeric vector of survival probabilities
}
\description{
Generate survival probabilities for a survival distribution
at the specified times.
}
\examples{
dist1 <- define_surv_param('exp', rate = 0.12)
surv_prob(dist1, c(0, 1, 2, 3))
  
}
